Move PersonProjectMemberships to org.zanata.service
